Fix legacy build errors.
