+++ politeiawww/identity.go
@@ -1,0 +1,52 @@
+// Copyright (c) 2017-2021 The Decred developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package main
+
+import (
+	"bufio"
+	"os"
+	"path/filepath"
+
+	"github.com/decred/politeia/util"
+)
+
+// getIdentity fetches the remote identity from politeiad.
+func getIdentity(rpcHost, rpcCert, rpcIdentityFile, interactive string) error {
+	id, err := util.RemoteIdentity(false, rpcHost, rpcCert)
+	if err != nil {
+		return err
+	}
+
+	// Pretty print identity.
+	log.Infof("Identity fetched from politeiad")
+	log.Infof("Key        : %x", id.Key)
+	log.Infof("Fingerprint: %v", id.Fingerprint())
+
+	if interactive != allowInteractive {
+		// Ask user if we like this identity
+		log.Infof("Press enter to save to %v or ctrl-c to abort",
+			rpcIdentityFile)
+		scanner := bufio.NewScanner(os.Stdin)
+		scanner.Scan()
+		if err = scanner.Err(); err != nil {
+			return err
+		}
+	} else {
+		log.Infof("Saving identity to %v", rpcIdentityFile)
+	}
+
+	// Save identity
+	err = os.MkdirAll(filepath.Dir(rpcIdentityFile), 0700)
+	if err != nil {
+		return err
+	}
+	err = id.SavePublicIdentity(rpcIdentityFile)
+	if err != nil {
+		return err
+	}
+	log.Infof("Identity saved to: %v", rpcIdentityFile)
+
+	return nil
+}
