+++ politeiawww/legacy/cmsevents.go
@@ -5,7 +5,7 @@
 package legacy
 
 import (
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 )
 
 const (
