+++ politeiawww/legacy/codetracker/github/database/cockroachdb/cockroachdb.go
@@ -11,7 +11,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/decred/politeia/politeiawww/codetracker/github/database"
+	"github.com/decred/politeia/politeiawww/legacy/codetracker/github/database"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/postgres"
 )
