+++ politeiawww/legacy/routes.go
@@ -8,7 +8,16 @@ import (
 	"net/http"
 	"strings"
 
+	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
+	cmv1 "github.com/decred/politeia/politeiawww/api/comments/v1"
+	piv1 "github.com/decred/politeia/politeiawww/api/pi/v1"
+	rcv1 "github.com/decred/politeia/politeiawww/api/records/v1"
+	tkv1 "github.com/decred/politeia/politeiawww/api/ticketvote/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
+	"github.com/decred/politeia/politeiawww/comments"
+	"github.com/decred/politeia/politeiawww/pi"
+	"github.com/decred/politeia/politeiawww/records"
+	"github.com/decred/politeia/politeiawww/ticketvote"
 	"github.com/decred/politeia/util"
 )
 
@@ -21,7 +30,7 @@ const (
 )
 
 // setUserWWWRoutes setsup the user routes.
-func (p *LegacyPoliteiawww) SetUserWWWRoutes() {
+func (p *LegacyPoliteiawww) setUserWWWRoutes() {
 	// Public routes
 	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
 		www.RouteNewUser, p.handleNewUser,
@@ -29,84 +38,81 @@ func (p *LegacyPoliteiawww) SetUserWWWRoutes() {
 	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
 		www.RouteVerifyNewUser, p.handleVerifyNewUser,
 		permissionPublic)
-	/*
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteResendVerification, p.handleResendVerification,
-			permissionPublic)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteLogout, p.handleLogout,
-			permissionPublic)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteResetPassword, p.handleResetPassword,
-			permissionPublic)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteVerifyResetPassword, p.handleVerifyResetPassword,
-			permissionPublic)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserDetails, p.handleUserDetails,
-			permissionPublic)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUsers, p.handleUsers,
-			permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteResendVerification, p.handleResendVerification,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteLogout, p.handleLogout,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteResetPassword, p.handleResetPassword,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteVerifyResetPassword, p.handleVerifyResetPassword,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserDetails, p.handleUserDetails,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUsers, p.handleUsers,
+		permissionPublic)
 
-		// Setup the login route.
-		p.addLoginRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteLogin, p.handleLogin)
+	// Setup the login route.
+	p.addLoginRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteLogin, p.handleLogin)
 
-		// Routes that require being logged in.
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteSecret, p.handleSecret,
-			permissionLogin)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserMe, p.handleMe,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteUpdateUserKey, p.handleUpdateUserKey,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteVerifyUpdateUserKey, p.handleVerifyUpdateUserKey,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteChangeUsername, p.handleChangeUsername,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteChangePassword, p.handleChangePassword,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteEditUser, p.handleEditUser,
-			permissionLogin)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserRegistrationPayment, p.handleUserRegistrationPayment,
-			permissionLogin)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserProposalPaywall, p.handleUserProposalPaywall,
-			permissionLogin)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserProposalPaywallTx, p.handleUserProposalPaywallTx,
-			permissionLogin)
-		p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
-			www.RouteUserProposalCredits, p.handleUserProposalCredits,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteSetTOTP, p.handleSetTOTP,
-			permissionLogin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteVerifyTOTP, p.handleVerifyTOTP,
-			permissionLogin)
+	// Routes that require being logged in.
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteSecret, p.handleSecret,
+		permissionLogin)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserMe, p.handleMe,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteUpdateUserKey, p.handleUpdateUserKey,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteVerifyUpdateUserKey, p.handleVerifyUpdateUserKey,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteChangeUsername, p.handleChangeUsername,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteChangePassword, p.handleChangePassword,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteEditUser, p.handleEditUser,
+		permissionLogin)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserRegistrationPayment, p.handleUserRegistrationPayment,
+		permissionLogin)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserProposalPaywall, p.handleUserProposalPaywall,
+		permissionLogin)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserProposalPaywallTx, p.handleUserProposalPaywallTx,
+		permissionLogin)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteUserProposalCredits, p.handleUserProposalCredits,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteSetTOTP, p.handleSetTOTP,
+		permissionLogin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteVerifyTOTP, p.handleVerifyTOTP,
+		permissionLogin)
 
-		// Routes that require being logged in as an admin user.
-		p.addRoute(http.MethodPut, www.PoliteiaWWWAPIRoute,
-			www.RouteUserPaymentsRescan, p.handleUserPaymentsRescan,
-			permissionAdmin)
-		p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
-			www.RouteManageUser, p.handleManageUser,
-			permissionAdmin)
-	*/
+	// Routes that require being logged in as an admin user.
+	p.addRoute(http.MethodPut, www.PoliteiaWWWAPIRoute,
+		www.RouteUserPaymentsRescan, p.handleUserPaymentsRescan,
+		permissionAdmin)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteManageUser, p.handleManageUser,
+		permissionAdmin)
 }
 
-/*
 // setCMSUserWWWRoutes setsup the user routes for cms mode
-func (p *LegacyPoliteiawwww) setCMSUserWWWRoutes() {
+func (p *LegacyPoliteiawww) setCMSUserWWWRoutes() {
 	// Public routes
 	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
 		www.RouteLogout, p.handleLogout,
@@ -176,9 +182,6 @@ func (p *LegacyPoliteiawwww) setCMSUserWWWRoutes() {
 }
 
 func (p *LegacyPoliteiawww) setCMSWWWRoutes() {
-	// Return a 404 when a route is not found
-	p.router.NotFoundHandler = http.HandlerFunc(p.handleNotFound)
-
 	// The version routes set the CSRF token and thus need to be part
 	// of the CSRF protected auth router.
 	p.auth.HandleFunc("/", p.handleVersion).Methods(http.MethodGet)
@@ -268,14 +271,16 @@ func (p *LegacyPoliteiawww) setCMSWWWRoutes() {
 		cms.RouteUserCodeStats, p.handleUserCodeStats,
 		permissionLogin)
 
-	// Unauthenticated websocket
-	p.addRoute("", www.PoliteiaWWWAPIRoute,
-		www.RouteUnauthenticatedWebSocket, p.handleUnauthenticatedWebsocket,
-		permissionPublic)
-	// Authenticated websocket
-	p.addRoute("", www.PoliteiaWWWAPIRoute,
-		www.RouteAuthenticatedWebSocket, p.handleAuthenticatedWebsocket,
-		permissionLogin)
+	/*
+		// Unauthenticated websocket
+		p.addRoute("", www.PoliteiaWWWAPIRoute,
+			www.RouteUnauthenticatedWebSocket, p.handleUnauthenticatedWebsocket,
+			permissionPublic)
+		// Authenticated websocket
+		p.addRoute("", www.PoliteiaWWWAPIRoute,
+			www.RouteAuthenticatedWebSocket, p.handleAuthenticatedWebsocket,
+			permissionLogin)
+	*/
 
 	// Routes that require being logged in as an admin user.
 	p.addRoute(http.MethodPost, cms.APIRoute,
@@ -309,7 +314,153 @@ func (p *LegacyPoliteiawww) setCMSWWWRoutes() {
 		cms.RouteProposalBillingDetails, p.handleProposalBillingDetails,
 		permissionAdmin)
 }
-*/
+
+// setupPiRoutes sets up the API routes for piwww mode.
+func (p *LegacyPoliteiawww) setPiRoutes(r *records.Records, c *comments.Comments, t *ticketvote.TicketVote, pic *pi.Pi) {
+	// The version routes set the CSRF token and thus need to be part
+	// of the CSRF protected auth router.
+	p.auth.HandleFunc("/", p.handleVersion).Methods(http.MethodGet)
+	p.auth.StrictSlash(true).
+		HandleFunc(www.PoliteiaWWWAPIRoute+www.RouteVersion, p.handleVersion).
+		Methods(http.MethodGet)
+
+	// Legacy www routes. These routes have been DEPRECATED. Support
+	// will be removed in a future release.
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RoutePolicy, p.handlePolicy,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteTokenInventory, p.handleTokenInventory,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteAllVetted, p.handleAllVetted,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteProposalDetails, p.handleProposalDetails,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteBatchProposals, p.handleBatchProposals,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteVoteStatus, p.handleVoteStatus,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteAllVoteStatus, p.handleAllVoteStatus,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteActiveVote, p.handleActiveVote,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteCastVotes, p.handleCastVotes,
+		permissionPublic)
+	p.addRoute(http.MethodGet, www.PoliteiaWWWAPIRoute,
+		www.RouteVoteResults, p.handleVoteResults,
+		permissionPublic)
+	p.addRoute(http.MethodPost, www.PoliteiaWWWAPIRoute,
+		www.RouteBatchVoteSummary, p.handleBatchVoteSummary,
+		permissionPublic)
+
+	// Record routes
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RoutePolicy, r.HandlePolicy,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteNew, r.HandleNew,
+		permissionLogin)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteEdit, r.HandleEdit,
+		permissionLogin)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteSetStatus, r.HandleSetStatus,
+		permissionAdmin)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteDetails, r.HandleDetails,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteTimestamps, r.HandleTimestamps,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteRecords, r.HandleRecords,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteInventory, r.HandleInventory,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteInventoryOrdered, r.HandleInventoryOrdered,
+		permissionPublic)
+	p.addRoute(http.MethodPost, rcv1.APIRoute,
+		rcv1.RouteUserRecords, r.HandleUserRecords,
+		permissionPublic)
+
+	// Comment routes
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RoutePolicy, c.HandlePolicy,
+		permissionPublic)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteNew, c.HandleNew,
+		permissionLogin)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteVote, c.HandleVote,
+		permissionLogin)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteDel, c.HandleDel,
+		permissionAdmin)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteCount, c.HandleCount,
+		permissionPublic)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteComments, c.HandleComments,
+		permissionPublic)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteVotes, c.HandleVotes,
+		permissionPublic)
+	p.addRoute(http.MethodPost, cmv1.APIRoute,
+		cmv1.RouteTimestamps, c.HandleTimestamps,
+		permissionPublic)
+
+	// Ticket vote routes
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RoutePolicy, t.HandlePolicy,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteAuthorize, t.HandleAuthorize,
+		permissionLogin)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteStart, t.HandleStart,
+		permissionAdmin)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteCastBallot, t.HandleCastBallot,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteDetails, t.HandleDetails,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteResults, t.HandleResults,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteSummaries, t.HandleSummaries,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteSubmissions, t.HandleSubmissions,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteInventory, t.HandleInventory,
+		permissionPublic)
+	p.addRoute(http.MethodPost, tkv1.APIRoute,
+		tkv1.RouteTimestamps, t.HandleTimestamps,
+		permissionPublic)
+
+	// Pi routes
+	p.addRoute(http.MethodPost, piv1.APIRoute,
+		piv1.RoutePolicy, pic.HandlePolicy,
+		permissionPublic)
+	p.addRoute(http.MethodPost, piv1.APIRoute,
+		piv1.RouteSetBillingStatus, pic.HandleSetBillingStatus,
+		permissionAdmin)
+	p.addRoute(http.MethodPost, piv1.APIRoute,
+		piv1.RouteSummaries, pic.HandleSummaries,
+		permissionPublic)
+}
 
 // addRoute sets up a handler for a specific method+route. If method is not
 // specified it adds a websocket.
@@ -343,6 +494,20 @@ func (p *LegacyPoliteiawww) addRoute(method string, routeVersion string, route s
 		// Add route to public router
 		p.router.StrictSlash(true).HandleFunc(fullRoute, handler).Methods(method)
 	}
+}
+
+// addLoginRoute sets up a handler for the login route. The login route is
+// special. It is the only public route that requires CSRF protection, so we
+// use a separate function to register it.
+func (p *LegacyPoliteiawww) addLoginRoute(method string, routeVersion string, route string, handler http.HandlerFunc) {
+	// Sanity check
+	if !strings.Contains(route, "login") {
+		panic("you cannot use this function to register non login routes")
+	}
+
+	// Add login route to the auth router
+	fullRoute := routeVersion + route
+	p.auth.StrictSlash(true).HandleFunc(fullRoute, handler).Methods(method)
 }
 
 // isLoggedIn ensures that a user is logged in before calling the next
