+++ politeiawww/legacy/www.go
@@ -7,43 +7,23 @@ package legacy
 import (
 	"encoding/hex"
 	"encoding/json"
-	"fmt"
 	"net/http"
-	"net/http/httputil"
 
 	"github.com/decred/politeia/politeiad/api/v1/mime"
+	v1 "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/util"
 	"github.com/decred/politeia/util/version"
 	"github.com/gorilla/csrf"
 )
-
-// handleNotFound is a generic handler for an invalid route.
-func (p *LegacyPoliteiawww) handleNotFound(w http.ResponseWriter, r *http.Request) {
-	// Log incoming connection
-	log.Debugf("Invalid route: %v %v %v %v",
-		util.RemoteAddr(r), r.Method, r.URL, r.Proto)
-
-	// Trace incoming request
-	log.Tracef("%v", newLogClosure(func() string {
-		trace, err := httputil.DumpRequest(r, true)
-		if err != nil {
-			trace = []byte(fmt.Sprintf("logging: "+
-				"DumpRequest %v", err))
-		}
-		return string(trace)
-	}))
-
-	util.RespondWithJSON(w, http.StatusNotFound, www.ErrorReply{})
-}
 
 // version is an HTTP GET to determine the lowest API route version that this
 // backend supports.  Additionally it is used to obtain a CSRF token.
 func (p *LegacyPoliteiawww) handleVersion(w http.ResponseWriter, r *http.Request) {
 	log.Tracef("handleVersion")
 
-	versionReply := www.VersionReply{
-		Version:      www.PoliteiaWWWAPIVersion,
-		Route:        www.PoliteiaWWWAPIRoute,
+	versionReply := v1.VersionReply{
+		Version:      v1.PoliteiaWWWAPIVersion,
+		Route:        v1.PoliteiaWWWAPIRoute,
 		BuildVersion: version.BuildMainVersion(),
 		PubKey:       hex.EncodeToString(p.cfg.Identity.Key[:]),
 		TestNet:      p.cfg.TestNet,
@@ -68,7 +48,7 @@ func (p *LegacyPoliteiawww) handleVersion(w http.ResponseWriter, r *http.Request
 	w.Header().Set("X-Frame-Options", "DENY")
 	w.Header().Set("X-XSS-Protection", "1; mode=block")
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	w.Header().Set(www.CsrfToken, csrf.Token(r))
+	w.Header().Set(v1.CsrfToken, csrf.Token(r))
 
 	w.WriteHeader(http.StatusOK)
 	w.Write(vr)
@@ -78,26 +58,26 @@ func (p *LegacyPoliteiawww) handlePolicy(w http.ResponseWriter, r *http.Request)
 	// Get the policy command.
 	log.Tracef("handlePolicy")
 
-	reply := &www.PolicyReply{
-		MinPasswordLength:          www.PolicyMinPasswordLength,
-		MinUsernameLength:          www.PolicyMinUsernameLength,
-		MaxUsernameLength:          www.PolicyMaxUsernameLength,
-		UsernameSupportedChars:     www.PolicyUsernameSupportedChars,
-		ProposalListPageSize:       www.ProposalListPageSize,
-		UserListPageSize:           www.UserListPageSize,
-		MaxImages:                  www.PolicyMaxImages,
-		MaxImageSize:               www.PolicyMaxImageSize,
-		MaxMDs:                     www.PolicyMaxMDs,
-		MaxMDSize:                  www.PolicyMaxMDSize,
+	reply := &v1.PolicyReply{
+		MinPasswordLength:          v1.PolicyMinPasswordLength,
+		MinUsernameLength:          v1.PolicyMinUsernameLength,
+		MaxUsernameLength:          v1.PolicyMaxUsernameLength,
+		UsernameSupportedChars:     v1.PolicyUsernameSupportedChars,
+		ProposalListPageSize:       v1.ProposalListPageSize,
+		UserListPageSize:           v1.UserListPageSize,
+		MaxImages:                  v1.PolicyMaxImages,
+		MaxImageSize:               v1.PolicyMaxImageSize,
+		MaxMDs:                     v1.PolicyMaxMDs,
+		MaxMDSize:                  v1.PolicyMaxMDSize,
 		PaywallEnabled:             p.paywallIsEnabled(),
 		ValidMIMETypes:             mime.ValidMimeTypes(),
-		MinProposalNameLength:      www.PolicyMinProposalNameLength,
-		MaxProposalNameLength:      www.PolicyMaxProposalNameLength,
-		ProposalNameSupportedChars: www.PolicyProposalNameSupportedChars,
-		MaxCommentLength:           www.PolicyMaxCommentLength,
-		TokenPrefixLength:          www.TokenPrefixLength,
+		MinProposalNameLength:      v1.PolicyMinProposalNameLength,
+		MaxProposalNameLength:      v1.PolicyMaxProposalNameLength,
+		ProposalNameSupportedChars: v1.PolicyProposalNameSupportedChars,
+		MaxCommentLength:           v1.PolicyMaxCommentLength,
+		TokenPrefixLength:          v1.TokenPrefixLength,
 		BuildInformation:           version.BuildInformation(),
-		IndexFilename:              www.PolicyIndexFilename,
+		IndexFilename:              v1.PolicyIndexFilename,
 		MinLinkByPeriod:            0,
 		MaxLinkByPeriod:            0,
 		MinVoteDuration:            0,
