+++ politeiawww/records/events.go
@@ -6,7 +6,7 @@ package records
 
 import (
 	v1 "github.com/decred/politeia/politeiawww/api/records/v1"
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 )
 
 const (
